move crafting classes
