+++ src/main/java/org/cloudburstmc/server/CloudServer.java
@@ -38,10 +38,10 @@ import org.cloudburstmc.server.config.CloudburstYaml;
 import org.cloudburstmc.server.config.ServerConfig;
 import org.cloudburstmc.server.config.ServerProperties;
 import org.cloudburstmc.server.console.NukkitConsole;
+import org.cloudburstmc.server.crafting.CraftingManager;
 import org.cloudburstmc.server.event.CloudEventManager;
 import org.cloudburstmc.server.inject.CloudburstModule;
 import org.cloudburstmc.server.inject.CloudburstPrivateModule;
-import org.cloudburstmc.server.inventory.CraftingManager;
 import org.cloudburstmc.server.level.*;
 import org.cloudburstmc.server.level.storage.StorageIds;
 import org.cloudburstmc.server.locale.LocaleManager;
