+++ src/main/java/org/cloudburstmc/server/blockentity/FurnaceBlockEntity.java
@@ -20,8 +20,8 @@ import org.cloudburstmc.api.item.ItemTypes;
 import org.cloudburstmc.api.item.data.Bucket;
 import org.cloudburstmc.api.level.chunk.Chunk;
 import org.cloudburstmc.api.util.Direction;
+import org.cloudburstmc.server.crafting.FurnaceRecipe;
 import org.cloudburstmc.server.inventory.CloudFurnaceInventory;
-import org.cloudburstmc.server.inventory.FurnaceRecipe;
 import org.cloudburstmc.server.item.ItemUtils;
 import org.cloudburstmc.server.player.CloudPlayer;
 import org.cloudburstmc.server.registry.CloudItemRegistry;
