+++ src/main/java/org/cloudburstmc/server/crafting/MixRecipe.java
@@ -1,4 +1,4 @@
-package org.cloudburstmc.server.inventory;
+package org.cloudburstmc.server.crafting;
 
 import lombok.ToString;
 import org.cloudburstmc.api.inventory.Recipe;
