+++ src/main/java/org/cloudburstmc/server/inventory/ContainerRecipe.java
@@ -4,6 +4,7 @@ import org.cloudburstmc.api.inventory.RecipeType;
 import org.cloudburstmc.api.item.ItemStack;
 import org.cloudburstmc.api.util.Identifier;
 import org.cloudburstmc.api.util.Identifiers;
+import org.cloudburstmc.server.crafting.MixRecipe;
 
 public class ContainerRecipe extends MixRecipe {
     public ContainerRecipe(Identifier id, ItemStack input, ItemStack ingredient, ItemStack output) {
