+++ src/main/java/org/cloudburstmc/server/registry/CloudRecipeRegistry.java
@@ -25,7 +25,8 @@ import org.cloudburstmc.api.registry.RegistryException;
 import org.cloudburstmc.api.util.Identifier;
 import org.cloudburstmc.api.util.Identifiers;
 import org.cloudburstmc.server.Bootstrap;
-import org.cloudburstmc.server.inventory.*;
+import org.cloudburstmc.server.crafting.*;
+import org.cloudburstmc.server.inventory.ContainerRecipe;
 import org.cloudburstmc.server.item.CloudItemStack;
 import org.cloudburstmc.server.item.ItemUtils;
 import org.cloudburstmc.server.item.RecipeItemStack;
